# docs/swagger.yaml
definitions:
...
  handlers.AmountRequest:
    properties:
      amount:
        type: number
    type: object
...
  handlers.KartBookingRequest:
    properties:
      booking_id:
        type: integer
      kart_id:
        type: integer
    type: object
...
  models.Kart:
    properties:
      id:
        type: integer
      kartModel:
        $ref: '#/definitions/models.KartModel'
      kartModelID:
        type: integer
      kartodrom:
        $ref: '#/definitions/models.Kartodrom'
      kartodromID:
        type: integer
      status:
        $ref: '#/definitions/models.KartStatus'
    type: object
  models.KartModel:
    properties:
      category:
        $ref: '#/definitions/models.KartModelStatus'
      desc:
        type: string
      id:
        type: integer
      maxHeight:
        type: integer
      maxSpeed:
        type: integer
      maxWeight:
        type: integer
      name:
        type: string
      pricePerMin:
        type: number
    type: object
  models.KartModelStatus:
    enum:
    - детский
    - обычный
    - спортивный
    - гоночный
    - электрический
    type: string
    x-enum-varnames:
    - KidKart
    - CommonKart
    - SportKart
    - RaceKart
    - ElectoKart
  models.KartStatus:
    enum:
    - Доступен
    - В использовании
    - Сломан
    - Недоступен
    type: string
    x-enum-varnames:
    - Available
    - InUse
    - Broken
    - InStopList
  models.Kartodrom:
    properties:
      city:
        type: string
      email:
        type: string
      id:
        type: integer
      latitude:
        type: number
      location:
        type: string
      longitude:
        type: number
      name:
        type: string
      phone:
        type: string
      schedules:
        items:
          $ref: '#/definitions/models.KartodromSchedule'
        type: array
    type: object
  models.KartodromSchedule:
    properties:
      closeTime:
        type: string
      dayOfWeek:
        type: integer
      id:
        type: integer
      kartodrom:
        $ref: '#/definitions/models.Kartodrom'
      kartodromID:
        type: integer
      openTime:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /kartbookings:
    post:
      parameters:
      - description: Данные бронирования карта
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/handlers.KartBookingRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Зарезервировать карт для бронирования
      tags:
      - kartbooking
  /kartbookings/{bookingId}/{kartId}/activate:
    post:
      parameters:
      - description: ID бронирования
        in: path
        name: bookingId
        required: true
        type: integer
      - description: ID карта
        in: path
        name: kartId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Активировать бронирование карта
      tags:
      - kartbooking
  /kartbookings/{bookingId}/{kartId}/cancel:
    post:
      parameters:
      - description: ID бронирования
        in: path
        name: bookingId
        required: true
        type: integer
      - description: ID карта
        in: path
        name: kartId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Отменить бронирование карта
      tags:
      - kartbooking
  /kartbookings/{bookingId}/{kartId}/finish:
    post:
      parameters:
      - description: ID бронирования
        in: path
        name: bookingId
        required: true
        type: integer
      - description: ID карта
        in: path
        name: kartId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Завершить бронирование карта
      tags:
      - kartbooking
...
  /kartodrom/{id}/karts:
    get:
      parameters:
      - description: ID картодрома
        in: path
        name: id
        required: true
        type: integer
      - description: Дата начала (2025-05-15T10:00:00)
        in: query
        name: start
        required: true
        type: string
      - description: Дата окончания (2025-05-15T11:00:00)
        in: query
        name: end
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Kart'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить доступные карты для бронирования
      tags:
      - kartbooking
...
  /user/{id}/balance:
    get:
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить текущий баланс пользователя
      tags:
      - payment
...
  /user/{id}/payments:
    get:
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить историю платежей пользователя
      tags:
      - payment
  /user/{id}/refill:
    post:
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Сумма пополнения
        in: body
        name: amount
        required: true
        schema:
          $ref: '#/definitions/handlers.AmountRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Пополнить баланс пользователя
      tags:
      - payment
  /user/{id}/refund:
    post:
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Сумма возврата
        in: body
        name: amount
        required: true
        schema:
          $ref: '#/definitions/handlers.AmountRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Сделать возврат средств пользователю
      tags:
      - payment
swagger: "2.0"